Changed all the commands to have the new Parameter Stack parameter calls

+++ src/slogo/Model/Commands/ConcreteCommands/Forward.java
@@ -33,7 +33,7 @@ public class Forward extends Command {
    */
   @Override
   public Double executeAndReturnValue() {
-    distanceToTravel = database.getParameterStack().pop().doubleValue();
+    distanceToTravel = database.getParameterStack().pop();
     turtleHeading = turtleObject.getTurtleHeading();
 
     distanceProportionY = Math.sin(Math.toRadians(turtleHeading));
